feat: create Runner.java class
The Runner.java class will pass the parameters, options, and values to
the actual Picocli command. ExitCodeGenerator ensures we
have the getExitCode function, which will return the exitcode after
the command is run.

The interface CommandLineRunner tells Spring we are running a command line application. It requires we have the run method. This method is automatically run after the Spring application context is loaded during the main method.

IFactory will inject PicocliSpringFactory allows picocli subcommands to have annotations using dependency injection.

CheckCommand will be the class of the actual command. The run method
creates a new command and executes it with the arguments it will receive.